% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bluesprings_waggle_ela}
\alias{bluesprings_waggle_ela}
\title{Blue Springs ELA Data}
\format{
\subsection{\code{bluesprings_waggle_ela}}{

A dataframe
}
}
\usage{
bluesprings_waggle_ela
}
\description{
Blue Springs ELA Data
}
\keyword{datasets}
