% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{klein_waggle_math}
\alias{klein_waggle_math}
\title{Klein Waggle Math Data}
\format{
\subsection{\code{klein_waggle_math}}{

A dataframe
}
}
\usage{
klein_waggle_math
}
\description{
Klein Waggle Math Data
}
\keyword{datasets}
